backend/requirements.txt
fastapi
uvicorn
langchain-groq
langgraph
boto3
python-dotenv
